requirements.txt
flask
azure-cognitiveservices-vision-computervision
msrest